Removed GetInput and the private inherited functions from Actor base class, private inheritance doesn't make sense.
Had to add a public SetPosition function in my alt player to be able to set the position after initialization.

Removed Model.cpp from my solution, since it no longer existed.

Had to revert my changes to MazeGame back to be same as remote/master, same with Testboard, which is the class I'm using to test out a different algorithm for the maze and the random algorithms.